add -t pid option to task
	Allos task to attach to a running thread.
	Hit CTRL-C to stop and print counts.
	Cleanup print_counts()

+++ perf_examples/task.c
@@ -28,6 +28,7 @@
 #include <unistd.h>
 #include <string.h>
 #include <stdarg.h>
+#include <signal.h>
 #include <sys/wait.h>
 #include <err.h>
 
@@ -39,9 +40,11 @@ typedef struct {
 	int group;
 	int print;
 	int pin;
+	pid_t pid;
 } options_t;
 
 static options_t options;
+static volatile int quit;
 
 int
 child(char **arg)
@@ -55,7 +58,7 @@ child(char **arg)
 }
 
 static void
-print_counts(perf_event_desc_t *fds, int num, int do_delta)
+print_counts(perf_event_desc_t *fds, int num)
 {
 	uint64_t values[3];
 	int i, ret;
@@ -82,7 +85,7 @@ print_counts(perf_event_desc_t *fds, int num, int do_delta)
 		fds[i].value = val = perf_scale(values);
 		ratio = perf_scale_ratio(values);
 
-		val = do_delta ? (val - fds[i].prev_value): val;
+		val = val - fds[i].prev_value;
 		if (ratio == 1.0)
 			printf("%20"PRIu64" %s\n", val, fds[i].name);
 		else
@@ -94,6 +97,10 @@ print_counts(perf_event_desc_t *fds, int num, int do_delta)
 	}
 }
 
+static void sig_handler(int n)
+{
+	quit = 1;
+}
 
 int
 parent(char **arg)
@@ -107,61 +114,67 @@ parent(char **arg)
 	if (pfm_initialize() != PFM_SUCCESS)
 		errx(1, "libpfm initialization failed");
 
-	ret = pipe(ready);
-	if (ret)
-		err(1, "cannot create pipe ready");
-
-	ret = pipe(go);
-	if (ret)
-		err(1, "cannot create pipe go");
-
 	num = perf_setup_list_events(options.events, &fds);
 	if (num < 1)
 		exit(1);
 
-	/*
-	 * Create the child task
-	 */
-	if ((pid=fork()) == -1)
-		err(1, "Cannot fork process");
+	pid = options.pid;
+	if (!pid) {
+		ret = pipe(ready);
+		if (ret)
+			err(1, "cannot create pipe ready");
 
-	/*
-	 * and launch the child code
-	 */
-	if (pid == 0) {
-		close(ready[0]);
-		close(go[1]);
+		ret = pipe(go);
+		if (ret)
+			err(1, "cannot create pipe go");
+
 
 		/*
-		 * let the parent know we exist
+		 * Create the child task
 		 */
-               close(ready[1]);
-               if (read(go[0], &buf, 1) == -1)
-                       err(1, "unable to read go_pipe");
+		if ((pid=fork()) == -1)
+			err(1, "Cannot fork process");
+
+		/*
+		 * and launch the child code
+		 */
+		if (pid == 0) {
+			close(ready[0]);
+			close(go[1]);
+
+			/*
+			 * let the parent know we exist
+			 */
+			close(ready[1]);
+			if (read(go[0], &buf, 1) == -1)
+				err(1, "unable to read go_pipe");
 
 
-		exit(child(arg));
+			exit(child(arg));
+		}
+
+		close(ready[1]);
+		close(go[0]);
+
+		if (read(ready[0], &buf, 1) == -1)
+			err(1, "unable to read child_ready_pipe");
+
+		close(ready[0]);
 	}
-
-	close(ready[1]);
-	close(go[0]);
-
-	if (read(ready[0], &buf, 1) == -1)
-               err(1, "unable to read child_ready_pipe");
-
-	close(ready[0]);
 
 	fds[0].fd = -1;
 	for(i=0; i < num; i++) {
 		/*
 		 * create leader disabled with enable_on-exec
 		 */
-		if (options.group) {
-			fds[i].hw.disabled = !i;
-			fds[i].hw.enable_on_exec = !i;
-		} else {
-			fds[i].hw.disabled = 1;
-			fds[i].hw.enable_on_exec = 1;
+		if (!options.pid) {
+			if (options.group) {
+				fds[i].hw.disabled = !i;
+				fds[i].hw.enable_on_exec = !i;
+			} else {
+				fds[i].hw.disabled = 1;
+				fds[i].hw.enable_on_exec = 1;
+			}
 		}
 
 		/* request timing information necessary for scaling counts */
@@ -180,16 +193,27 @@ parent(char **arg)
 		}
 	}	
 
-	close(go[1]);
+	if (!options.pid)
+		close(go[1]);
 
 	if (options.print) {
-		while(waitpid(pid, &status, WNOHANG) == 0) {
-			print_counts(fds, num, 1);
-			sleep(1);
+		if (!options.pid) {
+			while(waitpid(pid, &status, WNOHANG) == 0) {
+				sleep(1);
+				print_counts(fds, num);
+			}
+		} else {
+			while(quit == 0) {
+				sleep(1);
+				print_counts(fds, num);
+			}
 		}
 	} else {
-		waitpid(pid, &status, 0);
-		print_counts(fds, num, 0);
+		if (!options.pid)
+			waitpid(pid, &status, 0);
+		else
+			pause();
+		print_counts(fds, num);
 	}
 
 	for(i=0; i < num; i++)
@@ -199,19 +223,21 @@ parent(char **arg)
 	return 0;
 error:
 	free(fds);
-	kill(SIGKILL, pid);
+	if (!options.pid)
+		kill(SIGKILL, pid);
 	return -1;
 }
 
 static void
 usage(void)
 {
-	printf("usage: task [-h] [-i] [-g] [-p] [-P] [-e event1,event2,...] cmd\n"
+	printf("usage: task [-h] [-i] [-g] [-p] [-P] [-t pid] [-e event1,event2,...] cmd\n"
 		"-h\t\tget help\n"
 		"-i\t\tinherit across fork\n"
 		"-g\t\tgroup events\n"
 		"-p\t\tprint counts every second\n"
 		"-P\t\tpin events\n"
+		"-t pid\tmeasure existing pid\n"
 		"-e ev,ev\tlist of events to measure\n"
 		);
 }
@@ -221,7 +247,7 @@ main(int argc, char **argv)
 {
 	int c;
 
-	while ((c=getopt(argc, argv,"he:igpP")) != -1) {
+	while ((c=getopt(argc, argv,"he:igpPt:")) != -1) {
 		switch(c) {
 			case 'e':
 				options.events = optarg;
@@ -238,6 +264,9 @@ main(int argc, char **argv)
 			case 'i':
 				options.inherit = 1;
 				break;
+			case 't':
+				options.pid = atoi(optarg);
+				break;
 			case 'h':
 				usage();
 				exit(0);
@@ -248,8 +277,10 @@ main(int argc, char **argv)
 	if (!options.events)
 		options.events = "PERF_COUNT_HW_CPU_CYCLES,PERF_COUNT_HW_INSTRUCTIONS";
 
-	if (!argv[optind])
-		errx(1, "you must specify a command to execute\n");
+	if (!argv[optind] && !options.pid)
+		errx(1, "you must specify a command to execute or a thread to attach to\n");
 	
+	signal(SIGINT, sig_handler);
+
 	return parent(argv+optind);
 }
